+++ packages/at_client/lib/src/manager/monitor.dart
@@ -8,7 +8,6 @@ import 'package:at_chops/at_chops.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/preference/monitor_preference.dart';
 import 'package:at_client/src/response/default_response_parser.dart';
-import 'package:at_client/src/util/network_util.dart';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_lookup/at_lookup.dart';
 import 'package:at_utils/at_logger.dart';
@@ -53,8 +52,6 @@ class Monitor {
   late RemoteSecondary _remoteSecondary;
 
   final DefaultResponseParser _defaultResponseParser = DefaultResponseParser();
-
-  late MonitorConnectivityChecker _monitorConnectivityChecker;
 
   late MonitorOutboundConnectionFactory _monitorOutboundConnectionFactory;
 
@@ -128,7 +125,6 @@ class Monitor {
     MonitorPreference monitorPreference,
     Function retryCallBack, {
     RemoteSecondary? remoteSecondary,
-    MonitorConnectivityChecker? monitorConnectivityChecker,
     MonitorOutboundConnectionFactory? monitorOutboundConnectionFactory,
     Duration? monitorHeartbeatInterval,
     this.atChops,
@@ -144,8 +140,6 @@ class Monitor {
     _remoteSecondary = remoteSecondary ??
         RemoteSecondary(atSign, preference, atChops: atChops);
     _retryCallBack = retryCallBack;
-    _monitorConnectivityChecker =
-        monitorConnectivityChecker ?? MonitorConnectivityChecker();
     _monitorOutboundConnectionFactory =
         monitorOutboundConnectionFactory ?? MonitorOutboundConnectionFactory();
     _heartbeatInterval =
@@ -171,17 +165,21 @@ class Monitor {
       _lastNotificationTime = lastNotificationTime;
     }
     try {
-      await _checkConnectivity();
-
       //1. Get a new outbound connection dedicated to monitor verb.
       _monitorConnection = await _createNewConnection(
           _atSign, _preference.rootDomain, _preference.rootPort);
-      _monitorConnection!.getSocket().listen(_messageHandler, onDone: () {
-        _logger.info(
-            'socket.listen onDone called. Will destroy socket, set status stopped, call retryCallback');
-        _callCloseStopAndRetry();
-      }, onError: (error) {
-        _logger.warning('socket.listen onError called with: $error');
+      runZonedGuarded(() {
+        _monitorConnection!.getSocket().listen(_messageHandler, onDone: () {
+          _logger.info(
+              'socket.listen onDone called. Will destroy socket, set status stopped, call retryCallback');
+          _callCloseStopAndRetry();
+        }, onError: (error) {
+          _logger.warning('socket.listen onError called with: $error');
+          _handleError(error);
+        });
+      }, (Object error, StackTrace stackTrace) {
+        _logger.warning(
+            'runZonedGuarded received socket error $error - calling _handleError');
         _handleError(error);
       });
       await _authenticateConnection();
@@ -390,10 +388,6 @@ class Monitor {
     }
   }
 
-  Future<void> _checkConnectivity() async {
-    await _monitorConnectivityChecker.checkConnectivity(_remoteSecondary);
-  }
-
   /// Handles messages on the inbound client's connection.
   /// Closes the inbound connection in case of any error.
   /// Throw a [BufferOverFlowException] if buffer is unable to hold incoming data
@@ -452,16 +446,6 @@ class Monitor {
 }
 
 enum MonitorStatus { notStarted, started, stopped, errored }
-
-class MonitorConnectivityChecker {
-  Future<void> checkConnectivity(RemoteSecondary remoteSecondary) async {
-    if (!(await NetworkUtil().isNetworkAvailable())) {
-      throw AtConnectException(
-          'Monitor connectivity checker: Internet connection unavailable');
-    }
-    return;
-  }
-}
 
 class MonitorOutboundConnectionFactory {
   Future<OutboundConnection> createConnection(String secondaryUrl,