+++ packages/at_client/lib/src/service/notification_service_impl.dart
@@ -45,7 +45,6 @@ class NotificationServiceImpl
 
   late AtClient _atClient;
   Monitor? _monitor;
-  ConnectivityListener? _connectivityListener;
 
   late AtClientManager _atClientManager;
   AtClientValidation atClientValidation = AtClientValidation();
@@ -126,21 +125,6 @@ class NotificationServiceImpl
       await _startMonitor();
       _logger.finer(
           '${_atClient.getCurrentAtSign()} monitor status: ${_monitor?.getStatus()}');
-      if (_connectivityListener == null) {
-        _connectivityListener = ConnectivityListener();
-        // Note that this subscription is cancelled by stopAllSubscriptions(), so we don't need to worry
-        // about this subscription accidentally starting the monitor after it has been intentionally stopped,
-        // for example by switching atSigns
-        _connectivityListener!.subscribe().listen((isConnected) {
-          if (isConnected) {
-            _logger.finer(
-                '${_atClient.getCurrentAtSign()} starting monitor through connectivity listener event');
-            _startMonitor();
-          } else {
-            _logger.finer('lost network connectivity');
-          }
-        });
-      }
     } finally {
       _initializing = false;
     }
@@ -229,8 +213,6 @@ class NotificationServiceImpl
         'stopAllSubscriptions() called - setting monitorIsPaused to true');
     monitorIsPaused = true;
     _monitor?.stop();
-    _connectivityListener?.unSubscribe();
-    _connectivityListener = null;
     _streamListeners.forEach((regex, streamController) {
       if (!streamController.isClosed) () => streamController.close();
     });