+++ packages/at_client/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 3.0.64
+- Made ConnectivityListener configurable, and removed some unnecessary network 
+  availability checks
 - fix: wrap Monitor's call to `socket.listen()` in a runZonedGuarded block
 ## 3.0.63
 - fix: Fixed bug in AtRpc.sendRequest which was causing repeat sends of requests